requirements.txt
django
djangorestframework-simplejwt
pillow
rest_framework
django-environ